third_party/WebKit/LayoutTests/animations/responsive/stacked-visibility-animations-responsive.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="target"></div>
<script>
test(() => {
  var firstAnimation = target.animate([{visibility: 'hidden'}, {visibility: 'visible'}], 1);
  var secondAnimation = target.animate({visibility: 'collapse'}, 1);
  firstAnimation.pause();
  firstAnimation.currentTime = 0;
  secondAnimation.pause();
  secondAnimation.currentTime = 0.75;
  assert_equals(getComputedStyle(target).visibility, 'collapse');
  firstAnimation.currentTime = 0.1;
  assert_equals(getComputedStyle(target).visibility, 'visible');
}, "Visibility animations with neutral keyframes should be responsive to animations they're stacked on top of.");
</script>
